pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kenjhy <kenjhyx@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
requests = "^2.32.3"
pydantic = "^2.8.2"
uvicorn = "^0.30.6"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"
cachetools = "^5.5.0"
jinja2 = "^3.1.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
